+++ src/css/style.css
@@ -4,7 +4,7 @@
   /* colors */
   --primary-color: #f0a868;
   --secondary-color: #525b0f;
-  --tertiary-color: #a4b8c4;
+  --tertiary-color: #8a470c;
   --light-grey: #d0d0d0;
   --dark-grey: #303030;
 
@@ -13,23 +13,33 @@
   --small-font: 0.8em;
   --large-font: 1.2em;
 }
+
 * {
+  margin: 0;
+  padding: 0;
   box-sizing: border-box;
 }
+
 body {
-  margin: 0;
   font-family: var(--font-body);
   font-size: var(--font-base);
   color: var(--dark-grey);
 }
+
 img {
   max-width: 100%;
 }
+
 header {
   display: flex;
   justify-content: space-between;
-  padding: 0 10px;
+  padding: 0.5rem;
 }
+
+p {
+  padding: 0.75rem;
+}
+
 .logo {
   line-height: 60px;
   width: 245px;
@@ -39,23 +49,28 @@ header {
   text-align: right;
   font-family: var(--font-headline);
 }
+
 .logo img {
   width: 60px;
   height: 60px;
 }
+
 .logo a {
   text-decoration: none;
   color: var(--font-body);
 }
+
 .hero {
   position: relative;
 }
+
 .hero .logo {
   position: absolute;
   left: calc(50% - 60px);
-  bottom: -10px;
+  bottom: -15px;
   background-color: white;
 }
+
 .logo--square {
   font-size: var(--small-font);
   width: 110px;
@@ -68,32 +83,42 @@ header {
   align-items: center;
   line-height: 1;
 }
+
 .logo--square img {
   width: 90px;
   height: 90px;
   margin-top: -1.3em;
 }
+
 .logo--square > div {
   margin-top: -1.7em;
 }
+
 /* utility classes */
 .highlight {
   color: var(--tertiary-color);
+}
+
+.title {
+  font-size: 2.2rem;
+  font-variant: small-caps;
+  text-shadow: 0 0 .5em #777;
 }
 
 button {
   padding: 0.5em 2em;
   background-color: var(--secondary-color);
   color: white;
-  margin: auto;
   display: block;
   border: 0;
   font-size: var(--large-font);
   cursor: pointer;
 }
+
 .divider {
-  border-bottom: 2px solid var(--primary-color);
+  border-bottom: 1px solid var(--primary-color);
 }
+
 /* End utility classes */
 
 /* Cart icon styles */
@@ -106,6 +131,7 @@ button {
 .cart svg {
   width: 25px;
 }
+
 .cart:hover svg {
   fill: gray;
 }
@@ -113,35 +139,73 @@ button {
 /* End cart icon styles */
 
 .mission {
-  padding: 0 0.5em;
-  /* margin: 0.5em 0; */
+  position: absolute;
+  top: 1rem;
+  left: 0;
+  right: 0;
   line-height: 1.3;
-  max-width: 600px;
-  margin: auto;
+  width: 100%;
+  background-color: var(--dark-grey);
+  color: var(--light-grey);
+  opacity: 0.8;
 }
+
+.mission p {
+  margin: 0 auto;
+  width: 50%;
+  text-align: center;
+}
+
+/* Product Grid */
+.product-grid {
+  columns: 4;
+  column-gap: 1rem;
+  list-style-type: none;
+  text-align: center;;
+}
+
+.product-grid a {
+  text-decoration: none;
+  color: var(--dark-grey);
+  font-size: var(--small-font);
+}
+
+.product-grid img {
+  width: min(100px, 90%);
+}
+
+
 .products {
-  padding: 0.5em;
+  margin-top: 2rem;
+  padding: 1rem;
 }
+
+.products h2 {
+  margin-bottom: 1em;
+}
+
 .product-list {
   display: flex;
   flex-flow: row wrap;
-  justify-content: space-between;
   list-style-type: none;
-  padding: 0;
+  align-items: stretch;
 }
+
 .product-list a {
   text-decoration: none;
   color: var(--dark-grey);
 }
+
 /* Styles for product lists */
 .product-card {
-  flex: 1 1 45%;
+  flex: 1 1 100%;
   margin: 0.25em;
-  padding: 0.5em;
+  padding: 1em;
   border: 1px solid var(--light-grey);
   max-width: 250px;
+  text-align: center;
 }
-.product-card h2,
+
 .product-card h3,
 .product-card p {
   margin: 0.5em 0;
@@ -154,16 +218,39 @@ button {
 .card__name {
   font-size: 1em;
 }
+
 /* End product list card */
 
 .product-detail {
+  
+  width: max(500px, 90%);
+  margin: 1em auto;
   padding: 1em;
-  max-width: 500px;
-  margin: auto;
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  align-items: center;
+  grid-gap: 2em;
+}
+
+.product-detail h2 {
+  grid-column: 1/3;
+  padding: .5em;
+  text-align: center;
+  background-color: var(--dark-grey);
+  color: #fff;
 }
 
 .product-detail img {
   width: 100%;
+}
+
+.product-detail h2 {
+  margin-bottom: 0.5em;
+}
+
+#p-price {
+  font-size: 1.5em;
+  color: var(--secondary-color);
 }
 
 /* Start cart list card styles */
@@ -179,9 +266,11 @@ button {
   grid-row: 1/3;
   grid-column: 1;
 }
+
 .cart-card p {
   margin-top: 0;
 }
+
 .cart-card .card__name {
   grid-row: 1;
   grid-column: 2;
@@ -199,6 +288,7 @@ button {
 
 .cart-card__price {
   grid-row: 2;
+ 
   grid-column: 3;
 }
 
@@ -207,9 +297,11 @@ button {
     max-width: 1080px;
     margin: auto;
   }
+
   .mission {
     font-size: 1.2em;
   }
+
   .cart-card {
     font-size: inherit;
     grid-template-columns: 150px auto 15%;
@@ -218,5 +310,6 @@ button {
 
 footer {
   font-size: var(--small-font);
+  border-top: 1px solid var(--tertiary-color);
   padding: 1em;
 }
