+++ src/index.html
@@ -21,6 +21,27 @@
           <div>Sleep<span class="highlight">Outside</span></div>
         </div>
       </div>
+      <section class="product-categories">
+        <h2>Shop by Category</h2>
+        <div class="categories">
+          <a href="product_listing/index.html?category=tents" class="category">
+            <img src="images/tents.svg" alt="Tents">
+            <p>Tents</p>
+          </a>
+          <a href="product_listing/index.html?category=backpacks" class="category">
+            <img src="images/backpacks.svg" alt="Backpacks">
+            <p>Backpacks</p>
+          </a>
+          <a href="product_listing/index.html?category=sleeping-bags" class="category">
+            <img src="images/sleeping-bags.svg" alt="Sleeping Bags">
+            <p>Sleeping Bags</p>
+          </a>
+          <a href="product_listing/index.html?category=hammocks" class="category">
+            <img src="images/hammocks.svg" alt="Hammocks">
+            <p>Hammocks</p>
+          </a>
+        </div>
+      </section>
       <section class="mission">
         <p>We spend too much time these days indoors...sitting usually.</p>
         <p>No wonder we are a mess!</p>
