+++ src/js/ProductData.mjs
@@ -1,27 +1,27 @@
+const baseURL = import.meta.env.VITE_SERVER_URL;
+
 function convertToJson(res) {
   if (res.ok) {
     return res.json();
   } else {
-    throw new Error('Bad Response');
+    throw new Error("Bad Response");
   }
 }
 
 export default class ProductData {
-  constructor(category) {
-    this.category = category;
-    this.path = `../json/${this.category}.json`;
+  async getData(category) {
+    try {
+      const response = await fetch(`${baseURL}products/search/${category}`);
+      const data = await convertToJson(response);
+      return data.Result;
+    } catch (error) {
+      console.error("Error fetching product data:", error);
+      return []; // Return an empty array if fetch fails
+    }
   }
-  getData() {
-    return fetch(this.path)
-      .then(convertToJson)
-      .then((data) => data)
-      .catch ((error) => {
-        console.error("Error fetching product data:", error);
-        return []; // Return empty array if fetch fails
-      });
-  }
-  async findProductById(id) {
-    const products = await this.getData();
+
+  async findProductById(id, category) {
+    const products = await this.getData(category);
     return products.find((item) => item.Id === id);
   }
 }
