+++ src/js/ProductList.mjs
@@ -3,7 +3,7 @@ import { renderListWithTemplate } from "./utils.mjs";
 function productCardTemplate(product) {
   return `<li class="product-card">
     <a href="product_pages/index.html?product=${product.Id}">
-      <img src="${product.Image}" alt="Image of ${product.Name}" />
+      <img src="${product.ImageUrl}" alt="Image of ${product.Name}" />
     </a>
     <h3 class="card__brand">${product.Brand.Name}</h3>
     <h2 class="card__name">${product.Name}</h2>
@@ -20,8 +20,8 @@ export default class ProductList {
   }
 
   async init() {
-    this.fullList = await this.dataSource.getData(); // Store all products
-    this.renderList(this.fullList); // Render all products initially
+    this.fullList = await this.dataSource.getData(this.category); // Pass category
+    this.renderList(this.fullList); // Render products
   }
 
   renderList(list) {
